feat: Implement CustomListView for accurate drop position detection in file management
